requirements.txt
openai==1.9.0
openapi-schema-pydantic==1.2.4
streamlit==1.23.1
opencv-python-headless
fastapi==0.109.2
uvicorn==0.27.1